requirements.txt
numpy
django
django-crispy-forms
Pillow
mpi4py
django-cleanup
scipy
matplotlib
django-background-tasks